Simplify class-hierarchy of member-side Cache
In this commit none of the functionality is added or removed.

In order to simplify classes of Cache the following steps are done:
1. CacheProxy contains mostly all overrides from ICacheInternal. All methods are tend to be without
	deep low-level code.
2. CacheProxySupport is only one superclass of CacheProxy that contains internal re-used methods
	and low level code.
3. Inner class LoadAllTask is moved outside of CacheProxySupport to reduce complexity.
4. CountDownLatches and all related methods are moved to a standalone class
	CacheProxySyncListenerCompleter
	for logical isolation and complexity reasons.
5. Cache entry listener registrations and all related methods are moved to a standalone class
	CacheProxySyncListenerCompleter.
